Fixed numerous things
fixed numerous UI and inventory bugs such as the pickaxe not lighting up and being prompted to buy more than one boat. Also, the power bar now updates energy on screen instead of just behind the scenes. I'm still not sure if encountering an obstacle with the proper tool removes from inventory or visually decrements the tool or count. Not sure if we care. I commented the html, css, and game.js as well as frontEnd.js. I also updated the contribution sheet. Will do the overview / guide that Warren wants as part of the deliverables tomorrow.